FastAPI/requirements.txt
fastapi
uvicorn
sqlalchemy
starlette
email-validator
python-jose
passlib
bcrypt
python-dotenv
python-multipart
aiofiles
slowapi

# pip install -r requirements.txt
# '-r' reads the contents of requirements.txt